requirements.txt
GitPython==3.1.27
pygit2==1.9.1
PyGithub==1.55
gitflic==0.11
requests==2.27.1
click~=8.0.4